% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_election.R
\name{simulate_election}
\alias{simulate_election}
\title{Simulate an election}
\usage{
simulate_election(
  num_voters,
  num_districts,
  num_ballots,
  num_groups,
  num_candidates,
  num_attributes,
  lambda = 0.5,
  seed = 0
)
}
\arguments{
\item{num_voters}{amount of voters per table}

\item{num_districts}{number of districts}

\item{num_ballots}{number of tables}

\item{num_groups}{number of groups}

\item{num_candidates}{number of candidates}

\item{num_attributes}{number of attributes}

\item{lambda}{Shuffling fraction of voters (0=no shuffle, 1=full shuffle)}

\item{seed}{RNG seed}
}
\description{
Simulate an election
}
