% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip-class.R
\name{eip}
\alias{eip}
\title{S3 object for the Expectaction-Maximization Algorithm}
\usage{
eip(X = NULL, W = NULL, E = NULL, V = NULL, json_path = NULL)
}
\arguments{
\item{X}{A \verb{(b x c)} matrix representing candidate votes per ballot box.}

\item{W}{A \verb{(b x g)} matrix representing group votes per ballot box.}

\item{E}{A \verb{(b x d)} matrix representing the assignment of a ballot box towards a district.}

\item{V}{A \verb{(d x a)} matrix with the attributes for each district}

\item{json_path}{A path to a JSON file containing \code{X}, \code{W}, \code{E} and \code{V} fields, stored as nested arrays. It may contain additional fields with other attributes, which will be added to the returned object.}
}
\description{
S3 object for the Expectaction-Maximization Algorithm
}
